// src/lib/settings-types.ts
export interface SiteSettings {
  heroTitle: string
  heroSubtitle: string
  ctaPrimary: string
  ctaPrimaryLink: string
  ctaSecondary: string
  ctaSecondaryLink: string
  defaultTitle: string
  defaultDescription: string
  keywords: string
  ogImage: string
  logo: string
  favicon: string
  twitterUrl: string
  githubUrl: string
  linkedinUrl: string
  footerText: string
}